requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
python-dotenv==1.0.0
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21
WTForms==3.0.1
python-dateutil==2.8.2
Werkzeug==2.3.7
cryptography==41.0.4
Flask-RESTful==0.3.10
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
click==8.1.7
jinja2==3.1.2
Flask-Bootstrap==3.3.7.1
pytest==7.4.2
pytest-flask==1.2.0
flask-debugtoolbar==0.13.1
openpyxl==3.1.2
reportlab==4.0.7
